Draw each goroutine individually
Instead of collecting paths for all goroutines combined, do it per
goroutine. This increases the number of draw calls by a factor of the
number of visible goroutines, but it will allow us to encapsulate each
goroutine in its own widget, which simplifies offset calculations, event
handling, and will allow us to reuse goroutines elsewhere.