Improve handling of drags and tooltips
When panning the timeline, obtain a pointer grab, so that no child
widgets receive input. We don't want tooltips to appear and disappear as
the timeline gets panned.

For activity widgets, don't set hovered to true just because we're
dragging. Dragging can move off the activity, in which case tooltips
should disappear. Note that the panning grab is only set when panning,
not for other dragging (e.g. with the left mouse button), so activities
can still receive drag events.