trace: delta encode stacks
PCs tend to be somewhat close to each other, so delta encode stacks. If
we encounter a delta that's too large to represent, we fall back to just
using varints for PCs.

For our sample (a trace of running Staticcheck on std), this reduces the
size of all stacks by 67%. If we only used varints, the reduction would
be 52%, but that depends a lot on the concrete PCs, which is less of a
concern for the delta encoding.