trace: reduce peak memory usage
The old parser read all batches into memory, then merged them into a
single timeline. This required twice the memory, to hold the batches and
the merged result.

Our new parser interleaves parsing of events and merging into a
timeline. It does so using two passes. In the first pass, we only parse
enough of the trace to know where each batch begins. In the second pass,
we run the merging code, and load new batches on demand, whenever we've
consumed all events of a P. This works, because in order to merge
events, we only need the oldest unmerged event for each P.