trace: don't throw away goroutine information from GC-related events
The GC gets triggered by normal goroutines. This emits GC and STW
related events with ev.G set to the G that triggered the GC (or finished
it, in the case of the second STW phase). These events don't really
belong to a goroutine's main timeline (seeing a GCStart doesn't indicate
the end of a previous GoStart, for example), but it's still useful to
have this information, as it allows us to display which goroutine ended
up having to trigger GC.

We don't make use of this information yet, but this change also doesn't
break any existing code of ours.