Create ActivityWidgetTrack on demand
ActivityWidgetTrack is a very large struct, holding (currently) 28
instances of Gio's op.Ops, which alone amount to 3808 bytes. In large
traces we might have upward of a million tracks, making for almost 4 GB
of memory usage.

This state isn't needed unless the tracks are visible, so allocate them
lazily, by splitting the type in two: one tracking the spans and events,
and the other tracking the UI state. The former is expensive to compute
but small, while the latter is cheap to compute but big. We compute the
spans up front, and create the widgets when we need them, deallocating
them again once the widgets become hidden.

It's also worth noting that users rarely look at all goroutines in a
trace. Even when they scroll from top to bottom, either using the
scrollbar or by some animated navigation command, only a small fraction
of all goroutines will be visible, because when there are that many
goroutines, scrolling advances in increments of more than one goroutine.

We considered doing the same for ActivityWidget, but that type is much
smaller and it doesn't currently seem worth it.