Get rid of [5]byte event IDs
We've partially made the transition away from [5]byte in
f4bf26fb806589115a5216b8271db3e1ed8a2559. Complete it now. [5]byte is
awkward to use, especially across package boundaries. Its size was also
chosen for a specific layout of the Span type, which is prone to change.

We end up using a bit more memory now, but we can recover it in the future.