Clearly indicate which goroutines have returned
Add a zero-length span representing that a goroutine has returned. This
will render as a minimum width span in the timeline. We prevent this
span from merging with other spans so that it will always be visible,
regardless of zoom level.

We also improve goroutine tooltips, removing mentions of having returned
for goroutines that haven't in fact returned.

Closes: gh-65
Closes: gh-63