tinylfu: support generic keys
We unsafely convert non-string keys to strings so that we can hash them.
This is shallow and makes no effort to hash the contents of slices,
maps, strings (when the key itself isn't of type string) or any other
referenced data. This assumes that pointers do not change, that
identical strings have identical pointers, etc.