Cancel unused futures
Cancel futures that haven't been queried in a frame, under the
assumption that their results are no longer needed. Should that not be
the case, then the cancelled futures will restart themselves once
queried again.

Note that the only current use of futures, computing statistics, isn't
reacting to being cancelled. We may do it in the future, but for now,
it's not worth complicating the code.

This moves futures into the theme package for now.